# PAW-API_Producao/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Informação da RestAPI Produção",
        "version": "1.0.0",
        "title": "PAW - RestAPI Produção",
        "contact": {
            "name": "Grupo - Carlos Pinto, João Pires, Nuno Silva, Sérgio Rodrigues"
        }
    },
    "schemes": [
        "http"
    ],
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
        "/bomgroups": {
            "get": {
                "summary": "Obter todos os bomgroups",
                "description": "Obter todos os bomgroups",
                "produces": [
                    "application/json"
                ],
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "operação com sucesso",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Bomgroups"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            },
            "post": {
                "summary": "Criar novo bomgroup",
                "description": "Criar novo bomgroup",
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "bomgroup",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "type": "object",
                                "$ref": "#/definitions/Bomgroups"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Bomgroups"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
        "/bomgroups/bomgroup?page={nPaginas}&limit={limite}": {
            "get": {
                "summary": "Obter todos os bomgroups ordenados, definindo nº de páginas e limite",
                "description": "Obter todos os bomgroups ordenados, definindo nº de páginas e limite",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "nPaginas",
                        "description": "Número de páginas"
                    },
                    {
                        "in": "path",
                        "name": "limite",
                        "description": "Limite de grupos a apresentar"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Bomgroups"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
        "/bomgroups/{groupId}": {
            "get": {
                "summary": "Obter bomgroup por id",
                "description": "Obter bomgroup por id",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "groupId",
                        "description": "id do bomgroup a pesquisar",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "task": {
                                    "type": "object",
                                    "$ref": "#/definitions/Bomgroups"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Bomgroups"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            },
            "patch": {
                "summary": "Atualizar bomgroup",
                "description": "Atualizar bomgroup",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "groupId",
                        "in": "path",
                        "description": "id do bomgroup a alterar",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "item do bomgroup a alterar",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "type": "object",
                                "$ref": "#/definitions/Bomgroups"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Bomgroups"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            },
            "delete": {
                "summary": "Apagar bomgroup",
                "description": "Apagar bomgroup",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "groupId",
                        "in": "path",
                        "description": "id do bomgroup a eliminar",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Bomgroups"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
...
        "/auth/register": {
            "post": {
                "summary": "Criar novo utilizador",
                "description": "Criar novo utilizador",
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "novo utilizador",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "type": "object",
                                "$ref": "#/definitions/UserRegistration"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/UserRegistration"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "summary": "Iniciar sessão",
                "description": "Iniciar sessão",
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Iniciar sessão",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "type": "object",
                                "$ref": "#/definitions/UserLogin"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/UserLogin"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
...
        "/search/bomgroups?tags={querie}": {
            "get": {
                "summary": "Obter bomgroups através de pesquisa",
                "description": "Obter bomgroups através de pesquisa",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "querie",
                        "description": "termo a pesquisar",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "task": {
                                    "type": "object",
                                    "$ref": "#/definitions/Bomgroups"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Bomgroups"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "Bomgroups": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "cost": {
                    "type": "integer"
                },
                "notes": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "UserRegistration": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "role": {
                    "type": "object",
                    "properties": {
                        "name": {
                            "type": "string"
                        },
                        "scope": {
                            "type": "string",
                            "enum": [
                                "read-all",
                                "read-psts",
                                "manage-posts"
                            ]
                        }
                    }
                }
            }
        },
        "UserLogin": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "InvalidResponse": {
            "type": "object",
            "properties": {
                "statusCode": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}